エントリポイントのクラスを別ファイルに切り分け
