+++ src/lib/UI/MyApp.dart
@@ -1,0 +1,47 @@
+import 'package:flutter/material.dart';
+
+class MyApp extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+
+    String appTitle = 'App Memo';
+
+    return new MaterialApp(
+      title: appTitle,
+      theme: new ThemeData(
+        primarySwatch: Colors.orange,
+      ),
+      home: new MyHomePage(title: appTitle),
+    );
+  }
+}
+
+class MyHomePage extends StatefulWidget {
+  MyHomePage({Key key, this.title}) : super(key: key);
+  final String title;
+
+  @override
+  _MyHomePageState createState() => new _MyHomePageState();
+}
+
+class _MyHomePageState extends State<MyHomePage> {
+
+  @override
+  Widget build(BuildContext context) {
+    return new Scaffold(
+      appBar: new AppBar(
+        title: new Text(widget.title),
+      ),
+      body: new Center(
+        child: new Column(
+          mainAxisAlignment: MainAxisAlignment.center,
+          children: <Widget>[
+            new Text(
+              'Hello, World!',
+            ),
+          ],
+        ),
+      ),
+    );
+  }
+}
